hdr-metadata: Add WIP protocol for SMPTE ST2086/CTA-861.3 HDR metadata
TODO: Should we pass around the EOTF information that gets passed
to the display in the infoframe? It has provisions for using
tradiitonal gamma with HDR content. Need to figure out if we
should try to somehow allow clients to generate such content
and tell the compositor about it.

Also what's the story with changing the metadata (or even just the
colorspace) while the surface is alredy in use. If want to allow that
then we need to somehow guaranee that the chane happens on the correct
frame boundary. Attach the metadata/colorspace to a buffer instead?

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>